requirements.txt
# Core
langchain
langchain-core
langchain-community
langchain-openai
langgraph
pandas
openpyxl
python-dotenv
python-docx
pypdf
requests
tiktoken
certifi
arxiv
superlinked==19.21.1
sentence-transformers
mcp
fastmcp
